handle None type for get_file return

+++ functions/backup_restore/restore/restore_base.py
@@ -168,8 +168,9 @@ class RestoreBase:
                 except Exception as e:
                     self._handle_critical_failure(app_name, f"Failed to redeploy: {e}")
                     return False
+
             cnpg_delete_file = self.chart_info.get_file(app_name, "cnpg_pvcs_to_delete")
-            if cnpg_delete_file.exists():
+            if cnpg_delete_file and cnpg_delete_file.exists():
                 self._delete_cnpg_pvcs(cnpg_delete_file)
 
             return True
